+++ scripts/train_stage1.sh
@@ -42,6 +42,7 @@ torchrun --standalone --nproc_per_node=1 /path/to/tabicl/train/run.py \
             --icl_nhead 4 \
             --ff_factor 2 \
             --norm_first True \
+            --ignore_d False \
             --checkpoint_dir /my/stage1/checkpoint/dir \
             --save_temp_every 50 \
             --save_perm_every 5000
@@ -105,6 +106,7 @@ torchrun --standalone --nproc_per_node=1 /path/to/tabicl/train/run.py \
             --icl_nhead 4 \
             --ff_factor 2 \
             --norm_first True \
+            --ignore_d False \
             --checkpoint_dir /my/stage1/checkpoint/dir \
             --save_temp_every 50 \
             --save_perm_every 5000