+++ scripts/train_stage3.sh
@@ -47,6 +47,7 @@ torchrun --standalone --nproc_per_node=1 /path/to/tabicl/train/run.py \
             --icl_nhead 4 \
             --ff_factor 2 \
             --norm_first True \
+            --ignore_d False \
             --checkpoint_dir /my/stage3/checkpoint/dir \
             --checkpoint_path /my/stage2/checkpoint/dir/step-{latest}.ckpt \
             --save_temp_every 1 \
@@ -117,6 +118,7 @@ torchrun --standalone --nproc_per_node=1 /path/to/tabicl/train/run.py \
             --icl_nhead 4 \
             --ff_factor 2 \
             --norm_first True \
+            --ignore_d False \
             --checkpoint_dir /my/stage3/checkpoint/dir \
             --checkpoint_path /my/stage2/checkpoint/dir/step-{latest}.ckpt \
             --save_temp_every 1 \